Create SyncOU.ps1
This will take an Export of OU's form one forest and import them to another. The script will run but has some flaws, it can only go upto 15 nested OU's and it assumes a forest level Domain as your source. Would like to get the looping fixed and the ability to do both from CSV or direct from Forest to forests or Forest to child domain.